+++ integration-tests/cypress/cypress.config.js
@@ -1,13 +1,12 @@
 const { defineConfig } = require('cypress');
 const { registerSealightsTasks } = require('sealights-cypress-plugin');
+
 module.exports = defineConfig({
   e2e: {
     experimentalInteractiveRunEvents: true,
-    experimentalSessionAndOrigin: true ,
-    testIsolation: false ,
+    testIsolation: false,
     async setupNodeEvents(on, config) {
       await registerSealightsTasks(on, config);
-    },
+     },
   },
-
 });
