+++ integration-tests/cypress/cypress/support/commands.js
@@ -1,10 +1,8 @@
-// cypress/support/commands.js
-
 Cypress.Commands.add('apiRequest', (method, url, data = {}, options = {}) => {
-  return cy.request({
-    method,
-    url: Cypress.env('machine_dns') + url,
-    body: data,
-    ...options,
+    return cy.request({
+      method,
+      url: Cypress.env('machine_dns') + url,
+      body: data,
+      ...options,
+    });
   });
-});
