+++ integration-tests/cypress/cypress/support/e2e.js
@@ -1,1 +1,2 @@
+import './commands'
 import 'sealights-cypress-plugin/support';
